+++ turbo/apps/desktop/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.48.84](https://github.com/vm0-ai/okou/compare/desktop-v0.48.83...desktop-v0.48.84) (2026-09-17)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.459.0
+
 ## [0.48.83](https://github.com/vm0-ai/okou/compare/desktop-v0.48.82...desktop-v0.48.83) (2026-09-17)